ajout des authorisation via les policies et correction de l'url de la photo profil dans profile.vue(getProfilePhoto)

+++ app/Http/Controllers/API/UserController.php
@@ -103,8 +103,10 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
-       $user=User::findOrFail($id);
-       $user->delete();
+        $this->authorize('isAdmin');
+
+        $user=User::findOrFail($id);
+        $user->delete();
 
        return ['message'=>'user deleted'];
     }
